update website.html
Made a <div> element, taking me a while to figure out how to sepreate one div element from another. I'll figure it out soon enough though. My idea is to just put the css for one of the div in the html file. So something like
<div class = "divi"></div>
<style>
That's the best solution atm, but the code won't be as clean but again just a portfolio doesnt matter. This is my pet project.