making div element
Now, it looks like a top bar. The div element cannot be sepreated as it seems so what I'm probably going to be doing is just putting the stylesheet for one of the div elements into the html. Leave me any comments or pull requests as to how this might be different. Don't judge, I'm in 7th grade making websites and they aren't difficult but I just have to manage my time.